switch to using autogenerated tinker strings
