+++ Photon-Tinker/Base.lproj/TinkerStrings.strings
@@ -1,12 +1,12 @@
-/* 
-  TinkerStrings.strings
-  Particle
+// Generated on 2019-08-29 by iOS.py
 
-  Created by Raimundas Sakalauskas on 28/08/2019.
-  Copyright © 2019 Particle. All rights reserved.
-*/
+"TinkerStrings.Action.Cancel" = "Cancel123";
 
-"Tinker.Welcome.GetStartedButton" = "Get Started";
-"Tinker.Welcome.Version" = "v{{0}}b{{1}}";
+"TinkerStrings.DeviceList.SearchPlaceholder" = "Search devices...123";
+"TinkerStrings.DeviceList.Title" = "Your Devices123";
 
-"Tinker.DeviceList.Title" = "Your Devices123";
+"TinkerStrings.Welcome.GetStarted" = "Get Started123";
+"TinkerStrings.Welcome.Version" = "v{{0}}b{{1}}123";
+
+
+
